lustre: osc_cache: don't drop a lock we didn't take - two
osc_check_rpcs() drops a lock that it was called with, which can make
it harder to understand the code.

It is only called from one place, and that places takes the lock just
to call this function.
So instead, take the lock at the start of the function,
and drop it at the end.
This makes the code easier to follow.

Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: NeilBrown <neilb@suse.com>