lustre: embed typ_kobj in obd_type
As there is a 1-1 mapping between obd_types and their ->typ_kobj, it
is simple and more normal to embed the kobj in the obd_type, rather
than allocate it separately.

This requires calling "kobject_init()" earlier, so we open-code
relevant part of class_setup_tunables() in class_register_type().

A new simple_class_ktype is needed for the remaining user of
class_setup_tunables().

Signed-off-by: NeilBrown <neilb@suse.com>