lustre: osc: revert oe_users to atomic_t
A refcount_t is not supposed to ever be incremented
from zero - once it reaches zero it should stay there.
When building with CONFIG_REFCOUNT_FULL, an increment
from zero triggers a waring.

oe_users is allowed to be zero and in fact starts with
that value, so change it back to atomic_t.

Signed-off-by: NeilBrown <neilb@suse.com>