lustre: obdclass: fix incorrect LASSERT on lct_used
Asserting that lct_used is greater than 1 shortly before calling
atomic_dec_and_test() on is looks wrong.  We must expect it sometimes
to be 1, else the a_d_t will never succeed.

When the atomic_dec_and_test was introduced, the protocol for lct_used
was changed so that it gets decremented an extra time.  This assert
should have been changed then.

Fixes: ef84c0736421 ("staging: lustre: use wait_event_var() in lu_context_key_degister()")
Signed-off-by: NeilBrown <neilb@suse.com>