lnet: selftest: Use C99 struct initializer in framework.c
This patch makes no functional changes.  The struct initializer in
framework.c is updated to C99 syntax.

C89 positional initializers require values to be placed in the
correct order. This will cause errors if the fields of the struct
definition are reordered or fields are added or removed. C99 named
initializers avoid this problem, and also automatically clear any
values that are not explicitly set.

WC-bug-id: https://jira.whamcloud.com/browse/LU-6210
Lustre-commit: 60c9a2d351d4 ("LU-6210 lnet: Use C99 struct initializer in framework.c")
Signed-off-by: Steve Guminski <stephenx.guminski@intel.com>
Reviewed-on: https://review.whamcloud.com/28436
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>