lustre: llite: sync bdi sysfs name with lustre sysfs tree
When the linux kernel introduced super_setup_bdi_name()
the port for the lustre client was done which ends up
creating sysfs files of the format

/sys/devices/virtual/bdi/lustre-1

but their is not way to correlate this with the naming of
lustre sysfs objects. Update the bdi handle to use the naming
method applied to the llite sysfs tree.

WC-bug-id: https://jira.whamcloud.com/browse/LU-9558
Lustre-commit: 1defa62a19e7 ("LU-9558 llite: port lustre to unified handling of bdi")
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/28511
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>